Many differences from solution
I originally built my solution to do one frog and then was going to do a forEach method, but ran out of time so I just copy pasted.

In general there are a few places of duplication of code where I could instead call a function.  Just a matter of time management